// src/app/interfaces/manual.ts
export interface Manual {
    Id: number;
    Nombre: string;
    Incluye_dotes: boolean;
    Incluye_conjuros: boolean;
    Incluye_plantillas: boolean;
    Incluye_monstruos: boolean;
    Incluye_razas: boolean;
    Incluye_clases: boolean;
    Incluye_tipos: boolean;
    Incluye_subtipos: boolean;
    Oficial: boolean;
}